feat(dz65rgb:tkissing): move layer 2 functionality to userspace
