+++ users/tkissing/rules.mk
@@ -1,0 +1,1 @@
+SRC += tkissing.c
