+++ users/tkissing/tkissing.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "quantum.h"
+#include "action.h"
+#include "version.h"
+#include "rgblight_list.h"
+
+enum custom_keycodes {
+    KC_MAKE = SAFE_RANGE,
+    NEW_SAFE_RANGE  // use "NEW_SAFE_RANGE" for keymap specific codes
+};
